slightly restructured py dispatcher structure + bugfix
