+++ doc/dispatcher-guide.rst
@@ -1,0 +1,4 @@
+Writing a dispatcher
+====================
+
+gt
